249: intercept enemy FC
